+++ buildcontext/resolver.go
@@ -87,17 +87,24 @@ func NewResolver(cleanCollection *cleanup.Collection, gitLookup *GitLookup, cons
 // ExpandWildcard will expand a wildcard BUILD target in a local path or remote
 // Git repository. The pattern is the path relative to the target path and
 // should be in the form 'my/path/*'
-func (r *Resolver) ExpandWildcard(ctx context.Context, gwClient gwclient.Client, platr *platutil.Resolver, target domain.Target, pattern string) ([]string, error) {
-	var (
-		matches []string
-		err     error
-	)
+func (r *Resolver) ExpandWildcard(ctx context.Context, gwClient gwclient.Client, platr *platutil.Resolver, rootTarget, target domain.Target, pattern string) ([]string, error) {
 
 	if target.IsRemote() {
-		matches, err = r.gr.expandWildcard(ctx, gwClient, platr, target, pattern)
-	} else {
-		matches, err = fileutil.GlobDirs(target.GetLocalPath())
+		matches, err := r.gr.expandWildcard(ctx, gwClient, platr, target, pattern)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to expand remote BUILD target path")
+		}
+		return matches, nil
 	}
+
+	abs, err := filepath.Abs(rootTarget.LocalPath)
+	if err != nil {
+		return nil, errors.Wrapf(err, "could not determine absolute path for %q", rootTarget.LocalPath)
+	}
+
+	parent := filepath.Join(abs, target.GetLocalPath())
+
+	matches, err := fileutil.GlobDirs(parent)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to expand BUILD target path")
 	}